requirements.txt
streamlit
langgraph
langchain
langchain-openai
langchain-community
langchain-core
duckduckgo-search
requests
python-dotenv